chore(workflows): remove perf github summary
